Updated implementation based on adb commands
